# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ffmpeg-api
  version: 1.1.1
  summary: Async-first media processing API powered by FFmpeg
  description: >
    Draft OpenAPI contract for the next version of ffmpeg-api. This
    specification is intended to be the source of truth for Scalar
    documentation.
  license:
    name: ISC
    identifier: ISC
servers:
  - url: https://api.ffmpeg-api.dev
    description: Draft public base URL for production deployments.
...
security: []
paths:
  /v1/jobs/url:
    post:
      tags:
        - Jobs
      operationId: createJobFromUrl
      summary: Create a media processing job from a remote URL
      description: >
        Creates a processing job from a remote URL using application/json.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UrlJobCreateRequest'
            examples:
              oggFromUrl:
                summary: Speed up and trim silence from an OGG file
                value:
                  input:
                    type: url
                    url: https://cdn.example.com/audio/episode.ogg
                  analysis:
                    audioActivity: true
                  recipe:
                    output:
                      container: ogg
                      filename: episode-fast.ogg
                    operations:
                      - type: silence_trim
                        startThreshold: '-40dB'
                        startDuration: 0.3
                        stopThreshold: '-40dB'
                        stopDuration: 0.5
                      - type: speed
                        factor: 1.25
      responses:
        '200':
          description: Output generated synchronously
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Input too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Queue is full
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/jobs/upload:
    post:
      tags:
        - Jobs
      operationId: createJobFromUpload
      summary: Create a media processing job from an uploaded file
      description: >
        Creates a processing job from a multipart file upload.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadJobCreateRequest'
            examples:
              uploadJob:
                summary: Upload a file and process it
                value:
                  file: (binary)
                  mode: sync
                  analysis:
                    audioActivity: true
                  recipe:
                    output:
                      container: ogg
                      filename: upload-output.ogg
                    operations:
                      - type: speed
                        factor: 1.2
            encoding:
              recipe:
                contentType: application/json
              analysis:
                contentType: application/json
      responses:
        '200':
          description: Output generated synchronously
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Input too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Queue is full
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/jobs/{jobId}:
    get:
      tags:
        - Jobs
      operationId: getJob
      summary: Get job status
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/jobs/{jobId}/cancel:
    post:
      tags:
        - Jobs
      operationId: cancelJob
      summary: Cancel a queued or running job
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '202':
          description: Cancel request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Job can no longer be canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/jobs/{jobId}/artifact:
    get:
      tags:
        - Jobs
      operationId: downloadJobArtifact
      summary: Download the completed job artifact
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Generated artifact
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Artifact or job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Job is not completed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    UrlJobCreateRequest:
      type: object
      required:
        - input
        - recipe
      properties:
        input:
          type: object
          required:
            - type
            - url
          properties:
            type:
              type: string
              const: url
            url:
              type: string
              format: uri
        analysis:
          $ref: '#/components/schemas/JobAnalysisRequest'
        recipe:
          $ref: '#/components/schemas/TransformRecipe'
...
    UploadJobCreateRequest:
      type: object
      required:
        - file
        - recipe
      properties:
        file:
          type: string
          format: binary
        mode:
          type: string
          enum:
            - auto
            - sync
            - async
        analysis:
          description: Job analysis options sent as a JSON multipart part.
          $ref: '#/components/schemas/JobAnalysisRequest'
        recipe:
          description: Transform recipe sent as a JSON multipart part.
          $ref: '#/components/schemas/TransformRecipe'
...
    JobAnalysisRequest:
      type: object
      properties:
        audioActivity:
          oneOf:
            - type: boolean
            - $ref: '#/components/schemas/AudioActivityOptions'
    AudioActivityOptions:
      type: object
      properties:
        noiseThresholdDb:
          type: number
          example: -35
        minSilenceDuration:
          type: number
          example: 0.5
        minSegmentDuration:
          type: number
          example: 1.2
        minActiveRatio:
          type: number
          example: 0.08
        useSpeechBand:
          type: boolean
          example: true
    TransformRecipe:
      type: object
      required:
        - output
        - operations
      properties:
        output:
          type: object
          required:
            - container
          properties:
            container:
              type: string
              example: ogg
            filename:
              type: string
              example: episode-fast.ogg
        operations:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Operation'
...
    Operation:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - speed
            - silence_trim
            - trim
            - normalize
            - volume
            - channels
            - sample_rate
            - resize
            - fps
            - mute
            - extract_audio
            - watermark
      additionalProperties: true
    JobAccepted:
      type: object
      required:
        - id
        - status
        - links
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - queued
        links:
          type: object
          properties:
            self:
              type: string
              example: /v1/jobs/07d4a4a5-55f0-4c29-8790-b7546a9b2854
            artifact:
              type: string
              example: /v1/jobs/07d4a4a5-55f0-4c29-8790-b7546a9b2854/artifact
            cancel:
              type: string
              example: /v1/jobs/07d4a4a5-55f0-4c29-8790-b7546a9b2854/cancel
    Job:
      type: object
      required:
        - id
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - queued
            - downloading
            - processing
            - completed
            - failed
            - canceled
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        input:
          type: object
          additionalProperties: true
        recipe:
          $ref: '#/components/schemas/TransformRecipe'
        analysis:
          anyOf:
            - type: object
              properties:
                audioActivity:
                  $ref: '#/components/schemas/AudioActivityAnalysis'
            - type: 'null'
        progress:
          type: object
          properties:
            phase:
              type: string
              example: processing
            percent:
              type: number
              format: float
              minimum: 0
              maximum: 100
        artifact:
          anyOf:
            - type: object
              properties:
                filename:
                  type: string
                contentType:
                  type: string
                sizeBytes:
                  type: integer
                  format: int64
                downloadUrl:
                  type: string
                chunkCount:
                  anyOf:
                    - type: integer
                    - type: 'null'
            - type: 'null'
        metrics:
          type: object
          properties:
            queueWaitMs:
              type: integer
              format: int64
            downloadMs:
              type: integer
              format: int64
            processingMs:
              type: integer
              format: int64
            totalMs:
              type: integer
              format: int64
            inputBytes:
              type: integer
              format: int64
            outputBytes:
              type: integer
              format: int64
        error:
          $ref: '#/components/schemas/Error'
...
    AudioActivityAnalysis:
      type: object
      properties:
        analysisType:
          type: string
        warning:
          type: string
        settings:
          $ref: '#/components/schemas/AudioActivityOptions'
        durationSeconds:
          type: number
        silenceDurationSeconds:
          type: number
        activeDurationSeconds:
          type: number
        silenceRatio:
          type: number
        activeRatio:
          type: number
        longestActiveSegmentSeconds:
          type: number
        hasAudioActivity:
          type: boolean
        likelyBackgroundOnly:
          type: boolean
        likelyContainsSpeechLikeActivity:
          type: boolean
        silenceSegments:
          type: array
          items:
            $ref: '#/components/schemas/TimeSegment'
        activeSegments:
          type: array
          items:
            $ref: '#/components/schemas/TimeSegment'
    TimeSegment:
      type: object
      properties:
        start:
          type: number
        end:
          type: number
        duration:
          type: number
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: queue_full
        message:
          type: string
          example: The processing queue is full.
        details:
          anyOf:
            - type: object
              additionalProperties: true
            - type: 'null'